Add test infrastructure
This adds a docker container that's linked to a rabbitmq container. It
sets up the configuration for pigeon such that we have all the things we
need to run tests.

It adds a PigeonClient that lets us run tests. It's not
well-thought-out, but it is good enough for basic testing.

It adds a basic test that doesn't verify anything other than that things
didn't raise errors when we tried to do something.