Add tests for rabbitmq bits
This cleans up some of the rabbitmq code so it's easier to test and
clearer in its intentions. It matches the existing Socorro collector
rabbitmq behavior.

This creates a helper to make writing tests easier and the resulting
test code clearer.

Adds some tests verifying items did and did not get added to the queue
where appropriate.