Add monitoring and refactor handle()
This turns handle inside-out so that it first checks to see if it has any
crash_ids that need to be accepted and if it does, then creates a RabbitMQ
connection. Otherwise it doesn't.

This adds monitoring for accept, defer and pika_error counts using the format
that Datadog is expecting.

This adds a test for the accept case, too.